Working compile with nvcc of driver.cu
