--- build.bat
@@ -1,3 +1,0 @@
-mkdir cbuild
-cd cbuild
-cmake ..
