--- clean_and_build.bat
@@ -1,4 +1,0 @@
-rmdir /s /q cbuild
-mkdir cbuild
-cd cbuild
-cmake ..
